Organise item and resource type item transformers by namespace
- Organised transformers by namespace
