+++ app/Item/ResourceTypeItem/AllocatedExpense.php
@@ -74,7 +74,7 @@ class AllocatedExpense extends AbstractItem
      */
     public function transformer(array $data_to_transform): Transformer
     {
-        return new \App\Models\Transformers\ResourceTypeItemTypeAllocatedExpense($data_to_transform);
+        return new \App\Models\Transformers\ResourceTypeItemType\AllocatedExpense($data_to_transform);
     }
 
     /**
