+++ app/Item/SimpleExpense.php
@@ -173,7 +173,7 @@ class SimpleExpense extends AbstractItem
      */
     public function transformer(array $data_to_transform): Transformer
     {
-        return new \App\Models\Transformers\ItemTypeSimpleExpense($data_to_transform);
+        return new \App\Models\Transformers\ItemType\SimpleExpense($data_to_transform);
     }
 
     /**
