+++ app/Models/Transformers/ResourceTypeItemType/SimpleExpense.php
@@ -1,7 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models\Transformers;
+namespace App\Models\Transformers\ResourceTypeItemType;
+
+use App\Models\Transformers\Transformer;
 
 /**
  * Transform the data array into the format we require for the API
@@ -10,7 +12,7 @@ namespace App\Models\Transformers;
  * @copyright G3D Development Limited 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceTypeItemTypeAllocatedExpense extends Transformer
+class SimpleExpense extends Transformer
 {
     protected $item;
 
@@ -36,8 +38,6 @@ class ResourceTypeItemTypeAllocatedExpense extends Transformer
             'name' => $this->item['item_name'],
             'description' => $this->item['item_description'],
             'total' => number_format((float) $this->item['item_total'], 2, '.', ''),
-            'percentage' => (int) $this->item['item_percentage'],
-            'actualised_total' => number_format((float) $this->item['item_actualised_total'], 2, '.', ''),
             'effective_date' => $this->item['item_effective_date'],
             'created' => $this->item['item_created_at'],
             'resource' => [
